playout: move bent-four logic to tactics/
also noticed that previous code didn't board_play() passes,
fixed now.
