+++ playout.h
@@ -87,9 +87,6 @@ typedef struct {
 
 void amaf_init(amafmap_t *map);
 
-/* Init bent-four statics */
-void bent4_statics_init(int boardsize);
-
 /* Run one simulation and return score from white's perspective:
  *   >0: white wins
  *   <0: black wins
