+++ tactics/Makefile
@@ -1,5 +1,5 @@
 INCLUDES=-I..
-OBJS=dragon.o seki.o 1lib.o 2lib.o nlib.o ladder.o nakade.o selfatari.o util.o
+OBJS=dragon.o seki.o 1lib.o 2lib.o nlib.o ladder.o nakade.o selfatari.o bent4.o util.o
 
 all: lib.a
 lib.a: $(OBJS)
